Removed Typescript from example

+++ packages/perf/android/src/main/java/io/invertase/firebase/perf/UniversalFirebasePerfModule.java
@@ -117,15 +117,15 @@ public class UniversalFirebasePerfModule extends UniversalFirebaseModule {
       HttpMetric httpMetric = httpMetrics.get(id);
 
       if (httpMetricConfig.containsKey("httpResponseCode")) {
-        httpMetric.setHttpResponseCode((int) httpMetricConfig.getDouble("httpResponseCode"));
+        httpMetric.setHttpResponseCode(httpMetricConfig.getInt("httpResponseCode"));
       }
 
       if (httpMetricConfig.containsKey("requestPayloadSize")) {
-        httpMetric.setRequestPayloadSize((int) httpMetricConfig.getDouble("requestPayloadSize"));
+        httpMetric.setRequestPayloadSize(httpMetricConfig.getInt("requestPayloadSize"));
       }
 
       if (httpMetricConfig.containsKey("responsePayloadSize")) {
-        httpMetric.setResponsePayloadSize((int) httpMetricConfig.getDouble("responsePayloadSize"));
+        httpMetric.setResponsePayloadSize(httpMetricConfig.getInt("responsePayloadSize"));
       }
 
       if (httpMetricConfig.containsKey("responseContentType")) {
